Use Double instead of Int for internal BG representation in Sim Body

+++ sim-body/src/main/kotlin/de/dh/raaps/plugin/simbody/repository/SimBodyRepository.kt
@@ -13,7 +13,7 @@ class SimBodyRepository(private val simBodyDao: SimBodyDao) {
     suspend fun getSimulationState() = simBodyDao.getSimulationState()
 
     suspend fun updateSimulationState(
-        currentBgMgDl: Int,
+        currentBgMgDl: Double,
         lastTickTimestampMs: Long,
         exerciseIntensity: Double,
         stressLevel: Double,
